Fix PHP cross-version compatibility issue (PHP 7)
I'm a little aghast that nobody has come across this before and that no bug reports have come in about this (or maybe they have, but I haven't identified them as such), but the code as was was not cross-version compatible with PHP 5 vs PHP 7 due to the Uniform Variable Handling introduced in PHP 7.0.

A detailed code sample demonstrating the issue can be found here: https://3v4l.org/T3MDQ

The current fix maintains the pre-PHP 7.0 behaviour, which I believe, as this is an older codebase, is the intended behaviour.